move k8s-client-manager to pkg
