+++ internal/tools/monitor/core/diagnotor/controller/provider.go
@@ -25,7 +25,7 @@ import (
 	"github.com/erda-project/erda-infra/providers/httpserver"
 	"github.com/erda-project/erda-proto-go/core/monitor/diagnotor/pb"
 	"github.com/erda-project/erda/pkg/common/apis"
-	k8sclient "github.com/erda-project/erda/providers/k8s-client-manager"
+	k8sclient "github.com/erda-project/erda/pkg/k8s-client-manager"
 )
 
 type config struct {
